studygenie-backend/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
scikit-learn
python-multipart
python-dotenv
